+++ src/lib/st-sidebar/st-sidebar.component.ts
@@ -8,8 +8,10 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { ChangeDetectionStrategy, Component, Input, EventEmitter, Output, ChangeDetectorRef } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input, EventEmitter, Output } from '@angular/core';
+
 import { StSidebarItem } from './st-sidebar-item.interface';
+import { StSidebarVisualMode } from './st-sidebar-visual-mode';
 
 /**
  * @description {Component} [Sidebar]
@@ -45,13 +47,25 @@ export class StSidebarComponent {
    @Input() active: StSidebarItem;
    /** @Input {StSidebarItem[]} [items=''] List of items displayed on the menu */
    @Input() items: StSidebarItem[] = [];
+   /** @Input {boolean} [searchMode=''] Boolean to enable or disable the search mode. By default, it is disabled */
+   @Input() searchMode: boolean = false;
+   /** @Input {StSidebarVisualMode} [visualMode='StSidebarVisualMode.normal'] Visual mode used to display the item list */
+   @Input() visualMode: StSidebarVisualMode = StSidebarVisualMode.normal;
    /** @Output {StSidebarItem} [change=''] Event emitted when the active item is changed. This emits the active item */
    @Output() change: EventEmitter<StSidebarItem> = new EventEmitter<StSidebarItem>();
+   /** @Output {string} [search=''] Event emitted when search mode is enabled and user interacts with the search input */
+   @Output() search: EventEmitter<string> = new EventEmitter<string>();
 
 
    onChange(item: StSidebarItem): void {
       if (!this.active || this.active.id !== item.id) {
          this.change.emit(item);
+      }
+   }
+
+   onSearch(searchData: any): void {
+      if (searchData) {
+         this.search.emit(searchData.text);
       }
    }
 }